+++ agents-docs/navigation.ts
@@ -10,13 +10,27 @@ export default {
             'get-started/quick-start',
             'get-started/push-pull',
             'get-started/traces',
-            'get-started/credentials',
+            'get-started/upgrading',
           ],
         },
         'concepts',
       ],
     },
-
+    {
+      group: 'Tutorials',
+      pages: [
+        {
+          group: 'MCP Server Setup Guide',
+          icon: 'LuWrench',
+          pages: [
+            'tutorials/how-to-create-mcp-servers/overview',
+            'tutorials/how-to-create-mcp-servers/prebuilt-provider-hosted-mcps',
+            'tutorials/how-to-create-mcp-servers/prebuilt-composio-mcps',
+            'tutorials/how-to-create-mcp-servers/inkeep',
+          ],
+        },      
+      ],
+    },
     {
       group: 'Typescript SDK',
       icon: 'LuCode',
@@ -24,13 +38,22 @@ export default {
         'typescript-sdk/agent-settings',
         'typescript-sdk/agent-relationships',
         {
-          group: 'Tools and MCPs',
+          group: 'Tools',
           icon: 'LuHammer',
           pages: [
-            'typescript-sdk/tools/mcp-servers',
-            'typescript-sdk/tools/create-mcp-servers',
+            'typescript-sdk/tools/mcp-tools',
             'typescript-sdk/tools/function-tools',
-            'typescript-sdk/tools/credentials',
+          ],
+        },
+        {
+          group: 'Credentials',
+          icon: 'LuKey',
+          pages: [
+            'typescript-sdk/credentials/overview',
+            'typescript-sdk/credentials/nango',
+            'typescript-sdk/credentials/keychain',
+            'typescript-sdk/credentials/environment-variables',
+            'typescript-sdk/credentials/env-aware-credentials',
           ],
         },
         'typescript-sdk/headers',
@@ -108,10 +131,33 @@ export default {
           group: 'Chat Components',
           icon: 'LuBlocks',
           pages: [
-            'talk-to-your-agents/react/chat-button',
-            'talk-to-your-agents/react/custom-trigger',
-            'talk-to-your-agents/react/side-bar-chat',
-            'talk-to-your-agents/react/embedded-chat',
+            {
+              group: 'React',
+              icon: 'brand/ReactIcon',
+              pages: [
+                'talk-to-your-agents/react/chat-button',
+                'talk-to-your-agents/react/custom-trigger',
+                'talk-to-your-agents/react/side-bar-chat',
+                'talk-to-your-agents/react/embedded-chat',
+              ],
+            },
+            {
+              group: 'JavaScript',
+              icon: 'brand/JavascriptIcon',
+              pages: [
+                'talk-to-your-agents/javascript/chat-button',
+                'talk-to-your-agents/javascript/custom-trigger',
+                'talk-to-your-agents/javascript/side-bar-chat',
+                'talk-to-your-agents/javascript/embedded-chat',
+              ],
+            },
+            {
+              group: "Customization",
+              icon: "LuBrush",
+              pages: [
+                'ui-components/customization/styling',
+              ]
+            }
           ],
         },
         {